requirements.txt
Django==3.0
psycopg2-binary==2.8.5
djangorestframework==3.11.1
markdown==3.2.2
django-filter==2.3.0